+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+* Upgrade node to 18.x
+* Add scripts/differ.py
+* Fix lint make command
+
 # 2.1.4
 
 * Update plf sheet name to match latest spreadsheet template from AWS